scraper/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
BVCscrap==0.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
